requirements.txt
# Python 3.10.0
fastapi==0.115.0
pandas==2.2.3
numpy==1.26.4
faiss-cpu==1.8.0.post1
langchain-core==0.3.8
langchain-openai==0.2.1
langchain==0.3.2
dotenv==1.0.1
requests==2.32.3
spacy==3.8.2
beautifulsoup4==4.12.3
fr_core_news_sm @ https://github.com/explosion/spacy-models/releases/download/fr_core_news_sm-3.8.0/fr_core_news_sm-3.8.0-py3-none-any.whl#sha256=7d6ad14cd5078e53147bfbf70fb9d433c6a3865b695fda2657140bbc59a27e29
